Interactive Mouse Cursor Tracking and Click Simulation with OpenCV
In this the program loads an image and displays it in a window. As the mouse cursor moves over the image, its position is tracked and visualized in real-time by drawing a green circle at the cursor's location. Additionally, the script handles keyboard inputs for simulating left and right mouse clicks, as well as moving the cursor within the image window.